Moved files to namespace folders
